pyproject.toml
[tool.poetry]
name = "vsi2wm"
version = "0.1.0"
description = "VSI to WireMock converter - Convert CA LISA/Broadcom DevTest VSI files to WireMock stub mappings"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "vsi2wm"}]

[tool.poetry.dependencies]
python = "^3.11"
lxml = "^4.9.0"
pyyaml = "^6.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.5.0"

[tool.poetry.scripts]
vsi2wm = "vsi2wm.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
